Support PMCid gene lists in GMT input file
The Bader Lab added PMCid-associated gene lists to their monthly updated GMT file starting in December 2023. The loading of these GMT files and support for the lists formatting to ASCII from UTF-8 encoding in the case of some descriptions is now supported. PMC could be a seventh category of ontology for plotting Z score bar plots, but currently these PMCid entries are only in the table output when they meet criteria for minimum count of genes in the list, and at least some genes overlapping with the background from input unique genes.